backend/requirements.txt
fastapi
uvicorn
python-dotenv
langgraph
langgraph-checkpoint-mongodb
langchain-core
langchain-google-genai
pymongo
